Moved integration tests.
