+++ services/data_storage_service/src/database/diary_models.rs
@@ -568,31 +568,8 @@ async fn find_image_record(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{Config, Pool, app};
     use diesel::debug_query;
-    use diesel_async::pooled_connection::AsyncDieselConnectionManager;
-    use greenhouse_core::data_storage_service_dto::diary_dtos::{
-        get_diary::GetDiaryEntriesQueryDto, image_metadata::DiaryImageMetadataDto,
-    };
-    use reqwest::StatusCode;
-    use std::net::SocketAddr;
-    use testcontainers::ContainerAsync;
-    use testcontainers::ImageExt;
-    use testcontainers_modules::postgres::{self, Postgres};
-    use testcontainers_modules::testcontainers::runners::AsyncRunner;
-
-    struct TestApp {
-        base_url: String,
-        server: tokio::task::JoinHandle<()>,
-        container: ContainerAsync<Postgres>,
-    }
-
-    impl TestApp {
-        async fn shutdown(self) {
-            self.server.abort();
-            self.container.stop().await.unwrap();
-        }
-    }
+    use greenhouse_core::data_storage_service_dto::diary_dtos::image_metadata::DiaryImageMetadataDto;
 
     #[test]
     fn test_new_diary_entry() {
@@ -758,215 +735,5 @@ mod tests {
 
         assert!(sql.contains("storage_key"));
         assert!(!sql.contains("image_data"));
-    }
-
-    #[tokio::test]
-    async fn persists_tags_filters_entries_and_round_trips_images() {
-        let test_app = spawn_test_app().await;
-        let client = reqwest::Client::new();
-
-        let created = client
-            .post(format!("{}/diary/", test_app.base_url))
-            .json(&serde_json::json!({
-                "date": "2026-05-03T10:00:00Z",
-                "title": "Tomatoes",
-                "content": "Pruned and watered",
-                "tags": ["Harvest", " Tomatoes "]
-            }))
-            .send()
-            .await
-            .unwrap()
-            .json::<DiaryEntryResponseDto>()
-            .await
-            .unwrap();
-
-        assert_eq!(
-            created.tags,
-            vec![String::from("Harvest"), String::from("Tomatoes")]
-        );
-        assert!(created.images.is_empty());
-
-        client
-            .post(format!("{}/diary/", test_app.base_url))
-            .json(&serde_json::json!({
-                "date": "2026-05-03T11:00:00Z",
-                "title": "Cucumbers",
-                "content": "Checked growth",
-                "tags": ["Inspection"]
-            }))
-            .send()
-            .await
-            .unwrap();
-
-        let any_filter = client
-            .get(format!("{}/diary/", test_app.base_url))
-            .query(&GetDiaryEntriesQueryDto {
-                start: String::from("2026-05-03T00:00:00Z"),
-                end: String::from("2026-05-04T00:00:00Z"),
-                tags: vec![String::from("tomatoes")],
-                tag_filter_mode: DiaryTagFilterModeDto::Any,
-            })
-            .send()
-            .await
-            .unwrap()
-            .json::<greenhouse_core::data_storage_service_dto::diary_dtos::get_diary::GetDiaryResponseDto>()
-            .await
-            .unwrap();
-        assert_eq!(any_filter.entries.len(), 1);
-        assert_eq!(any_filter.entries[0].id, created.id);
-
-        let all_filter = client
-            .get(format!("{}/diary/", test_app.base_url))
-            .query(&GetDiaryEntriesQueryDto {
-                start: String::from("2026-05-03T00:00:00Z"),
-                end: String::from("2026-05-04T00:00:00Z"),
-                tags: vec![String::from("harvest"), String::from("tomatoes")],
-                tag_filter_mode: DiaryTagFilterModeDto::All,
-            })
-            .send()
-            .await
-            .unwrap()
-            .json::<greenhouse_core::data_storage_service_dto::diary_dtos::get_diary::GetDiaryResponseDto>()
-            .await
-            .unwrap();
-        assert_eq!(all_filter.entries.len(), 1);
-        assert_eq!(all_filter.entries[0].id, created.id);
-
-        let image_bytes = vec![137, 80, 78, 71];
-        let uploaded = client
-            .post(format!("{}/diary/{}/images", test_app.base_url, created.id))
-            .header("x-file-name", "leaf.png")
-            .header(reqwest::header::CONTENT_TYPE, "image/png")
-            .body(image_bytes.clone())
-            .send()
-            .await
-            .unwrap();
-        assert_eq!(uploaded.status(), StatusCode::OK);
-        let uploaded_metadata = uploaded.json::<DiaryImageMetadataDto>().await.unwrap();
-        assert_eq!(uploaded_metadata.file_name, "leaf.png");
-        assert_eq!(uploaded_metadata.media_type, "image/png");
-        assert_eq!(uploaded_metadata.byte_size, image_bytes.len() as i64);
-
-        let fetched_entry = client
-            .get(format!("{}/diary/{}", test_app.base_url, created.id))
-            .send()
-            .await
-            .unwrap()
-            .json::<DiaryEntryResponseDto>()
-            .await
-            .unwrap();
-        assert_eq!(fetched_entry.images, vec![uploaded_metadata.clone()]);
-
-        let downloaded = client
-            .get(format!(
-                "{}/diary/{}/images/{}",
-                test_app.base_url, created.id, uploaded_metadata.id
-            ))
-            .send()
-            .await
-            .unwrap();
-        assert_eq!(downloaded.status(), StatusCode::OK);
-        assert_eq!(
-            downloaded.headers()[reqwest::header::CONTENT_TYPE],
-            reqwest::header::HeaderValue::from_static("image/png")
-        );
-        assert_eq!(downloaded.bytes().await.unwrap().to_vec(), image_bytes);
-
-        let deleted = client
-            .delete(format!(
-                "{}/diary/{}/images/{}",
-                test_app.base_url, created.id, uploaded_metadata.id
-            ))
-            .send()
-            .await
-            .unwrap();
-        assert_eq!(deleted.status(), StatusCode::NO_CONTENT);
-
-        let after_delete = client
-            .get(format!("{}/diary/{}", test_app.base_url, created.id))
-            .send()
-            .await
-            .unwrap()
-            .json::<DiaryEntryResponseDto>()
-            .await
-            .unwrap();
-        assert!(after_delete.images.is_empty());
-
-        test_app.shutdown().await;
-    }
-
-    #[tokio::test]
-    async fn image_routes_return_not_found_for_missing_resources() {
-        let test_app = spawn_test_app().await;
-        let client = reqwest::Client::new();
-        let entry_id = Uuid::new_v4();
-        let image_id = Uuid::new_v4();
-
-        let missing_upload = client
-            .post(format!("{}/diary/{entry_id}/images", test_app.base_url))
-            .header("x-file-name", "missing.png")
-            .header(reqwest::header::CONTENT_TYPE, "image/png")
-            .body(vec![1, 2, 3])
-            .send()
-            .await
-            .unwrap();
-        assert_eq!(missing_upload.status(), StatusCode::NOT_FOUND);
-
-        let missing_download = client
-            .get(format!(
-                "{}/diary/{entry_id}/images/{image_id}",
-                test_app.base_url
-            ))
-            .send()
-            .await
-            .unwrap();
-        assert_eq!(missing_download.status(), StatusCode::NOT_FOUND);
-
-        let missing_delete = client
-            .delete(format!(
-                "{}/diary/{entry_id}/images/{image_id}",
-                test_app.base_url
-            ))
-            .send()
-            .await
-            .unwrap();
-        assert_eq!(missing_delete.status(), StatusCode::NOT_FOUND);
-
-        test_app.shutdown().await;
-    }
-
-    async fn spawn_test_app() -> TestApp {
-        let container = postgres::Postgres::default()
-            .with_db_name("data")
-            .with_tag("latest")
-            .start()
-            .await
-            .unwrap();
-        let db_port = container.get_host_port_ipv4(5432).await.unwrap();
-        let database_url = format!("postgres://postgres:postgres@localhost:{db_port}/data");
-
-        let config = Config {
-            service_port: 0,
-            database_url: database_url.clone(),
-            sentry_url: String::new(),
-            environment: String::from("test"),
-        };
-        let pool = Pool::builder()
-            .build(AsyncDieselConnectionManager::new(database_url))
-            .await
-            .unwrap();
-        let app = app(config, pool);
-
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0").await.unwrap();
-        let address: SocketAddr = listener.local_addr().unwrap();
-        let server = tokio::spawn(async move {
-            axum::serve(listener, app).await.unwrap();
-        });
-
-        TestApp {
-            base_url: format!("http://{address}"),
-            server,
-            container,
-        }
     }
 }
